Backend/requirements.txt
assemblyai
fastapi==0.110.2
google-generativeai
google-auth
gspread
numpy
opencv-python
Pillow==10.3.0
pydantic==2.7.1
pydub==0.25.1
pyodbc
python-dotenv
python-jose[cryptography]==3.3.0
python-multipart
reportlab==4.0.4
requests
SpeechRecognition==3.10.1
SQLAlchemy==2.0.28
uvicorn==0.29.0
